prov/verbs: Cleanup XRCD domain file if no longer in use
The XRC domain file is shared among processes on a system, and required
to be visible until all domains in all processes have opened the domain.
This code attempts to cleanup the file when domains are closed to avoid
conflicts when the same domain name is used with different permissions.
This solution is Linux specific.

Signed-off-by: Steve Welch <swelch@systemfabricworks.com>