prov/rxm: Rename several functions in rxm_cq
There are several rxm function with the prefix rxm_cq_, which have nothing
to do with the cq.  Remove the 'cq' from the prefix to make it clearer that
these are not cq related functions, and are merely handling received data.

Rename other calls to follow the more common format of 'verb-object' used
throughout the code.  The provider has a mix of function names, making it
difficult to figure out what action is being taken on what object.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>