core/getinfo: Allow util provider to request a specific core provider
When fi_getinfo is called, the hints passed by the user are checked
against a list of fi_info structures maintained by a util provider.
If a specific fi_info cannot meet the hints, it is skipped.  The
fi_info structures that pass are updated based on the hints and passed
to an fi_getinfo call to the core providers.

Currently, an fi_info structure maintained by the util provider cannot
be restricted to a specific core provider.  For example, rxm cannot have
an fi_info optimized for just the tcp provider.

To allow this, we need to expand the checks in ofi_check_fabric_attr to
include checks against the provider name.  This will allow skipping
fi_info structs that don't meet the application hints.  The check
against the provider name is in the main fi_getinfo call, which is how
the filtering works today.  But that check assumes that util providers
do not request specific core providers.

Once that check is in place, we need to copy the provider name from the
util provider fi_info into the hints passed to the core, via fi_getinfo.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>