core: Use base fi_info attributes when setting core fi_info hints
When calling fi_getinfo, the app hints are compared against a list
of fi_info structures maintained by a provider.  If a suitable fi_info
is found, the hints are accepted and processing continues.  If the
provider is a util provider, this involves creating hints for the core
provider and calling fi_getinfo recursively.

However, not all of the fi_info structs in the provider's list may
match the application hints.  So, the util provider does not know
which base fi_info matched, and must set all core hints generically.

Update the internal calls to pass the base fi_info struct, so that
it can be used to create the core hints.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>