prov/verbs: Implement XRC MSG EP as simplex data transfer EP
An XRC MSG EP is currently implemented to support duplex
data transfer. Since an XRC QP connection is from an initiator
QP to a target QP (simplex data transfer), when a connection is
requested the provider currently makes two connections (the
requested one plus one in the opposite direction). This commit
changes the provider to no longer make the second connection
by taking advantage of RxM support for simplex data transfer EP.
This simplifies the XRC connection setup logic, making the code
easier to maintain and in certain scenarios reducing the number
of connections that must be made.

Signed-off-by: Steve Welch <swelch@systemfabricworks.com>