prov/rxm: Allow forcing duplex connectivity with simplex transports
Allow an application to optionally force duplex connectivity when a
simplex connection is being established. This allows an application that
knows it requires duplex connectivity to establish that connectivity
before it is required. Normally following an accept, the duplex
connectivity is established when an RxM software protocol (e.g.
rendezvous or emulated atomics) requires a response/data be
returned to initiator or the application initiates a TX operation.

Signed-off-by: Steve Welch <swelch@systemfabricworks.com>